Converted EuiStepsHorizontal to TS
